feat: rename v5 upgrade to v6
